UI Code Cleanup + Reimplemented some missing features
Reorganized some of the code to hopefully increase readability and whatnot.

Reimplemented file copying and deleting. But there aren't any Yes/No confirm dialogs yet. I'll add those in a bit.

Still a ways to go. Next up is the settings menu, along with a rework of button input code.