More UI changes + collision code changes
Bunch of UI changes I procrastinated committing.

Also this branch is suppose to be dedicated to UI changes, but I got sidetracked and did a bunch of collision code tweaks, and I'm too lazy to sort stuff. But oh well, the branch will get merged eventually anyway.

Oh also the LHC extension was completely removed. The only thing I was using was its functions for doing asset tagging stuffs, but I found a (I think) better alternative to that.