Deceptively large commit
Updated to latest GM
-IDE v2024.13.1.193
-Runtime v2024.13.1.242

Made the move to Linux semi-recently, and had to start using the Steam version of GM. In doing so, a lot of things auto-formatted in some way. Mostly corrected filename cases but also a bunch of whitespace changes? Idk. (Maybe a lot of it only shows up in the GitHub app but will disappear once the commit is made.)

Made a lot of tweaks in a lot of places that I'm forgetting. Also you can now aim+shoot during the slide animation.